move remix-core code & tests to remix-debug
