+++ .gitignore
@@ -8,3 +8,4 @@ babelify-src
 docs/_build
 package-lock.json
 .DS_Store
+.tern-port
