--- remix-core/test/tests.js
@@ -1,3 +1,0 @@
-require('./traceManager.js')
-require('./codeManager.js')
-require('./disassembler.js')
