+++ remix-debug/index.js
@@ -2,6 +2,12 @@
 var remixCore = require('remix-core')
 var EthDebugger = require('./src/Ethdebugger')
 
+var CodeManager = require('./src/code/codeManager')
+var BreakpointManager = require('./src/code/breakpointManager')
+var StorageViewer = require('./src/storage/storageViewer')
+var StorageResolver = require('./src/storage/storageResolver')
+var TraceManager = require('./src/trace/traceManager')
+
 /*
   Use of breakPointManager :
 
@@ -12,11 +18,23 @@ var EthDebugger = require('./src/Ethdebugger')
 */
 module.exports = {
   EthDebugger: EthDebugger,
-   /**
-    * constructor
-    *
-    * @param {Object} _debugger - type of EthDebugger
-    * @return {Function} _locationToRowConverter - function implemented by editor which return a column/line position for a char source location
-    */
-  BreakpointManager: remixCore.code.BreakpointManager
+  /**
+   * constructor
+   *
+   * @param {Object} _debugger - type of EthDebugger
+   * @return {Function} _locationToRowConverter - function implemented by editor which return a column/line position for a char source location
+   */
+  BreakpointManager: remixCore.code.BreakpointManager,
+  code: {
+    CodeManager: CodeManager,
+    BreakpointManager: BreakpointManager
+  },
+  storage: {
+    StorageViewer: StorageViewer,
+    StorageResolver: StorageResolver
+  },
+  trace: {
+    TraceManager: TraceManager
+  }
 }
+
